requirements.txt
numpy~=1.0
torch
torchvision
wandb
hydra-core
pyyaml
omegaconf
colorama
tqdm
rich
opencv-python
matplotlib
lightning
click
h5py
moviepy
imageio
einops>=0.7
wandb-osh
diffusers
internetarchive
pytorchvideo
scipy
rotary_embedding_torch
torchmetrics[image]==0.11.4
tfrecord[torch]
bezier
timm
zmq
openai-clip
pyiqa==0.1.10
easydict
zmq
pytubefix
roma
transformers
